Fix implicit conversion of returned value in AssetLibrary::numGeometries()

+++ framework/include/minko/file/AssetLibrary.hpp
@@ -92,7 +92,7 @@ namespace minko
             uint
             numGeometries()
             {
-                return _geometries.size();
+                return static_cast<uint>(_geometries.size());
             }
 
             inline
